+++ sanpy/interface/plugins/__init__.py
@@ -40,6 +40,7 @@ from sanpy.interface.bScatterPlotWidget2 import myStatListWidget
 from .plotScatter import getPlotMarkersAndColors
 
 from .plotFi import plotFi
+from .nicepool_plugin import NicePoolPlugin
 
 # TODO: make this just one line, so user can drop a bplugin in and restart
 from . import *